+++ crates/ruff_linter/resources/test/fixtures/flake8_bandit/S704_extend_markup_names.py
@@ -2,5 +2,5 @@ from markupsafe import Markup
 from webhelpers.html import literal
 
 content = "<script>alert('Hello, world!')</script>"
-Markup(f"unsafe {content}")  # RUF035
-literal(f"unsafe {content}")  # RUF035
+Markup(f"unsafe {content}")  # S704
+literal(f"unsafe {content}")  # S704
