+++ crates/ruff_linter/resources/test/fixtures/flake8_bandit/S704_whitelisted_markup_calls.py
@@ -6,4 +6,4 @@ Markup(clean(content))
 
 # indirect assignments are currently not supported
 cleaned = clean(content)
-Markup(cleaned)  # RUF035
+Markup(cleaned)  # S704
