+++ crates/ruff_linter/src/rules/ruff/mod.rs
@@ -113,8 +113,6 @@ mod tests {
             &LinterSettings {
                 ruff: super::settings::Settings {
                     parenthesize_tuple_in_subscript: true,
-                    extend_markup_names: vec![],
-                    allowed_markup_calls: vec![],
                 },
                 ..LinterSettings::for_rule(Rule::IncorrectlyParenthesizedTupleInSubscript)
             },
@@ -130,8 +128,6 @@ mod tests {
             &LinterSettings {
                 ruff: super::settings::Settings {
                     parenthesize_tuple_in_subscript: false,
-                    extend_markup_names: vec![],
-                    allowed_markup_calls: vec![],
                 },
                 unresolved_target_version: PythonVersion::PY310,
                 ..LinterSettings::for_rule(Rule::IncorrectlyParenthesizedTupleInSubscript)
@@ -423,7 +419,6 @@ mod tests {
         Ok(())
     }
 
-    #[test_case(Rule::UnsafeMarkupUse, Path::new("RUF035.py"))]
     #[test_case(Rule::MapIntVersionParsing, Path::new("RUF048.py"))]
     #[test_case(Rule::MapIntVersionParsing, Path::new("RUF048_1.py"))]
     #[test_case(Rule::UnrawRePattern, Path::new("RUF039.py"))]
@@ -451,53 +446,6 @@ mod tests {
             &settings::LinterSettings {
                 preview: PreviewMode::Enabled,
                 ..settings::LinterSettings::for_rule(rule_code)
-            },
-        )?;
-        assert_messages!(snapshot, diagnostics);
-        Ok(())
-    }
-
-    #[test_case(Rule::UnsafeMarkupUse, Path::new("RUF035_extend_markup_names.py"))]
-    #[test_case(Rule::UnsafeMarkupUse, Path::new("RUF035_skip_early_out.py"))]
-    fn extend_allowed_callable(rule_code: Rule, path: &Path) -> Result<()> {
-        let snapshot = format!(
-            "extend_allow_callables__{}_{}",
-            rule_code.noqa_code(),
-            path.to_string_lossy()
-        );
-        let diagnostics = test_path(
-            Path::new("ruff").join(path).as_path(),
-            &LinterSettings {
-                ruff: super::settings::Settings {
-                    parenthesize_tuple_in_subscript: true,
-                    extend_markup_names: vec!["webhelpers.html.literal".to_string()],
-                    allowed_markup_calls: vec![],
-                },
-                preview: PreviewMode::Enabled,
-                ..LinterSettings::for_rule(rule_code)
-            },
-        )?;
-        assert_messages!(snapshot, diagnostics);
-        Ok(())
-    }
-
-    #[test_case(Rule::UnsafeMarkupUse, Path::new("RUF035_whitelisted_markup_calls.py"))]
-    fn whitelisted_markup_calls(rule_code: Rule, path: &Path) -> Result<()> {
-        let snapshot = format!(
-            "whitelisted_markup_calls__{}_{}",
-            rule_code.noqa_code(),
-            path.to_string_lossy()
-        );
-        let diagnostics = test_path(
-            Path::new("ruff").join(path).as_path(),
-            &LinterSettings {
-                ruff: super::settings::Settings {
-                    parenthesize_tuple_in_subscript: true,
-                    extend_markup_names: vec![],
-                    allowed_markup_calls: vec!["bleach.clean".to_string()],
-                },
-                preview: PreviewMode::Enabled,
-                ..LinterSettings::for_rule(rule_code)
             },
         )?;
         assert_messages!(snapshot, diagnostics);
