+++ crates/ruff_linter/src/rules/ruff/settings.rs
@@ -7,8 +7,6 @@ use std::fmt;
 #[derive(Debug, Clone, CacheKey, Default)]
 pub struct Settings {
     pub parenthesize_tuple_in_subscript: bool,
-    pub extend_markup_names: Vec<String>,
-    pub allowed_markup_calls: Vec<String>,
 }
 
 impl fmt::Display for Settings {
@@ -18,8 +16,6 @@ impl fmt::Display for Settings {
             namespace = "linter.ruff",
             fields = [
                 self.parenthesize_tuple_in_subscript,
-                self.extend_markup_names | array,
-                self.allowed_markup_calls | array,
             ]
         }
         Ok(())
